+++ examples/todo-app/src/api/todo.ts
@@ -40,10 +40,6 @@ export type Todo = {
    * @remarks Todo status
    */
   completed: boolean
-  /**
-   * @remarks Todo create time
-   */
-  createAt?: string | null | undefined
 }
 
 /**