+++ examples/todo-app/src/main.tsx
@@ -1,7 +1,13 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import { apiPipeline } from 'farrow-api-client'
 import './index.css'
 import App from './App'
+
+apiPipeline.use((request, next) => {
+  console.log('request', request)
+  return next()
+})
 
 ReactDOM.render(
   <React.StrictMode>