+++ packages/farrow-cors/CHANGELOG.md
@@ -1,5 +1,17 @@
 # farrow-cors
 
+## 2.0.0
+
+### Major Changes
+
+- 1f9b7fb: upgrade for version 2.0
+
+### Patch Changes
+
+- Updated dependencies [1f9b7fb]
+  - farrow-http@2.0.0
+  - farrow-pipeline@2.0.0
+
 ## 1.12.0
 
 ### Minor Changes