+++ packages/farrow-federation/CHANGELOG.md
@@ -1,5 +1,19 @@
 # farrow-federation
 
+## 2.0.0
+
+### Major Changes
+
+- 1f9b7fb: upgrade for version 2.0
+
+### Patch Changes
+
+- Updated dependencies [1f9b7fb]
+  - farrow-api@2.0.0
+  - farrow-api-server@2.0.0
+  - farrow-http@2.0.0
+  - farrow-schema@2.0.0
+
 ## 1.12.0
 
 ### Minor Changes