+++ packages/farrow-json-schema/CHANGELOG.md
@@ -1,5 +1,16 @@
 # farrow-json-schema
 
+## 2.0.0
+
+### Major Changes
+
+- 1f9b7fb: upgrade for version 2.0
+
+### Patch Changes
+
+- Updated dependencies [1f9b7fb]
+  - farrow-schema@2.0.0
+
 ## 1.12.0
 
 ### Minor Changes